@@ -1,13 +1,9 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { marker } from '@biesbjerg/ngx-translate-extract-marker';
 
 import { HomeComponent } from './home.component';
 
-const routes: Routes = [
-  { path: '', redirectTo: '/home', pathMatch: 'full' },
-  { path: 'home', component: HomeComponent },
-];
+const routes: Routes = [{ path: '', component: HomeComponent, title: 'klesiewicz.dev | home' }];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
